Deprecated LegacyRouteController and renamed PortableRouteController to RouteController
